+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.1.1 (unreleased)
 
 - Added `take` to vectors
+- Added `sample` to data frames
 
 ## 0.1.0 (2020-05-13)